backend/scraperjob/requirements.txt
beautifulsoup4>=4.12.2
requests>=2.31.0
lxml>=4.9.3
selenium>=4.11.2
python-dotenv>=1.0.0
urllib3>=2.0.4
click>=8.1.8
aiohttp>=3.11.11
extruct>=0.18.0
python-dateutil>=2.9.0
aiolimiter>=1.2.1
pyminiscraper>=0.1.1
asyncclick>=8.1.8.0
